Deeper clean on custom_arg_parser.py module.
A more-in-depth pass over a single module.

	- Add docstrings. (Enable help(function or class) and <object>.__doc__)
	- Add unit tests. (This will make future changes safer.)
	- Add optional static typing. (It really is worth the work. Catches some hard-to-debug errors.)
	- Mark private functions private (standard "_<name>" naming convention.)
	- Smooth out minor complaints from pylint.
	- Add TODO items that I now also think should be done.

Passes pep257 (docstrings standard), pytest (unit tests), mypy (optional static typing), pylint (lots of nit-picky things).

There's more work to be done (TODO) and more work on some or all items above that could be done,
but this is enough to move forward to the next module.